requirements.txt

django
djangorestframework
celery
redis
PyPDF2
scikit-learn
sentence-transformers
